requirements.txt
flask==2.0.1
google-cloud-texttospeech==2.12.0
google-generativeai==0.3.0
python-dotenv==0.19.0